# autorepro/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AutoRepro Enterprise",
    "description": "AI-powered Bug Operations SaaS Platform",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v1/bugs": {
      "post": {
        "tags": [
          "bugs"
        ],
        "summary": "Create Bug",
        "description": "Create a new bug report.",
        "operationId": "create_bug_api_v1_bugs_post",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BugCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "bugs"
        ],
        "summary": "List Bugs",
        "operationId": "list_bugs_api_v1_bugs_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Max items to return",
              "default": 20,
              "title": "Limit"
            },
            "description": "Max items to return"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Items to skip",
              "default": 0,
              "title": "Offset"
            },
            "description": "Items to skip"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/BugStatus"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by bug lifecycle status",
              "title": "Status"
            },
            "description": "Filter by bug lifecycle status"
          },
          {
            "name": "severity",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/BugSeverity"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by severity",
              "title": "Severity"
            },
            "description": "Filter by severity"
          },
          {
            "name": "assigned_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by assigned user UUID",
              "title": "Assigned To"
            },
            "description": "Filter by assigned user UUID"
          },
          {
            "name": "created_by",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by reporter UUID",
              "title": "Created By"
            },
            "description": "Filter by reporter UUID"
          },
          {
            "name": "team_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by team UUID",
              "title": "Team Id"
            },
            "description": "Filter by team UUID"
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/bugs/bulk/assign": {
      "post": {
        "tags": [
          "bulk"
        ],
        "summary": "Bulk Assign Endpoint",
        "description": "Assign multiple bugs to one user (SUPERVISOR+).",
        "operationId": "bulk_assign_endpoint_api_v1_bugs_bulk_assign_post",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkAssignRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/bugs/bulk/status": {
      "post": {
        "tags": [
          "bulk"
        ],
        "summary": "Bulk Update Status Endpoint",
        "description": "Update status for multiple bugs (SUPERVISOR+).",
        "operationId": "bulk_update_status_endpoint_api_v1_bugs_bulk_status_post",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkStatusRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/bugs/bulk/delete": {
      "post": {
        "tags": [
          "bulk"
        ],
        "summary": "Bulk Delete Endpoint",
        "description": "Soft delete multiple bugs (MANAGER+).",
        "operationId": "bulk_delete_endpoint_api_v1_bugs_bulk_delete_post",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkDeleteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BugCreate": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "target_url": {
            "type": "string",
            "title": "Target Url"
          },
          "priority": {
            "$ref": "#/components/schemas/BugPriority",
            "default": "MEDIUM"
          },
          "severity": {
            "$ref": "#/components/schemas/BugSeverity",
            "default": "MEDIUM"
          },
          "environment": {
            "type": "string",
            "title": "Environment",
            "default": "production"
          },
          "browser": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Browser"
          },
          "os": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Os"
          },
          "device": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device"
          },
          "team_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Id"
          }
        },
        "type": "object",
        "required": [
          "title",
          "description",
          "target_url"
        ],
        "title": "BugCreate"
      },
...
      "BugPriority": {
        "type": "string",
        "enum": [
          "LOW",
          "MEDIUM",
          "HIGH",
          "CRITICAL"
        ],
        "title": "BugPriority"
      },
      "BugSeverity": {
        "type": "string",
        "enum": [
          "LOW",
          "MEDIUM",
          "HIGH",
          "CRITICAL"
        ],
        "title": "BugSeverity"
      },
      "BugStatus": {
        "type": "string",
        "enum": [
          "CREATED",
          "TRIAGED",
          "ASSIGNED",
          "IN_PROGRESS",
          "RUNNING_AUTOREPRO",
          "RESOLVED",
          "CLOSED",
          "DUPLICATE"
        ],
        "title": "BugStatus",
        "description": "Bug lifecycle states \u2014 strict linear progression.\nRUNNING_AUTOREPRO is set by the worker when execution starts;\nroutes must not allow direct transition into it (only via job trigger)."
      },
...
      "BulkAssignRequest": {
        "properties": {
          "bug_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Bug Ids"
          },
          "assigned_to_user_id": {
            "type": "string",
            "format": "uuid",
            "title": "Assigned To User Id"
          }
        },
        "type": "object",
        "required": [
          "bug_ids",
          "assigned_to_user_id"
        ],
        "title": "BulkAssignRequest"
      },
      "BulkDeleteRequest": {
        "properties": {
          "bug_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Bug Ids"
          }
        },
        "type": "object",
        "required": [
          "bug_ids"
        ],
        "title": "BulkDeleteRequest"
      },
      "BulkStatusRequest": {
        "properties": {
          "bug_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Bug Ids"
          },
          "new_status": {
            "$ref": "#/components/schemas/BugStatus"
          }
        },
        "type": "object",
        "required": [
          "bug_ids",
          "new_status"
        ],
        "title": "BulkStatusRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}